Move fetching movies to home component
